Consolidating directories
Removed "examples" and moved to templates
