+++ templates/spotter-inject.cs
