+++ templates/spotter-process.cs
